Adding some stuff...
I finished of writing the get_config_string which replaces the get_string. On top of that I also wrote the set_config_string which lets you come to the conclusion that the File_Loader class is probably finished which is indeed so.

Apart from that I added a small method called test_method() which will be run if the pyhton file is ran. This helps us, although it is a module, to test all other methods in this Pyhton-File as they are called from there.