Reconfiguration of Files
In this patch the Client.py was renamed to Base.py. But instead of loosing the Client.py file I thought it would be a great idea to make an own base for all clients. That's why this file still exists.
On top of that I wrote a basic Database Controller which will allow easy accesss on all Database related things without any need to initialize and to login and so on.