Network support added
The most important thing I added in this patch was the Network support. It may cause some errors in the listening() method but I already have an idea to fix this, too.
I also put the whole network-traffic to different threads to ensure perfect and stable network communication without stopping the main-thread. The modules which use client or base have to watch, that they use the threads correctly and also delete them out of running_threads[] when they stop.
The last thing I added was a small skelleton for the Encrypter and Decrypter but as I said it is only a skelleton. (This Class will be in the base class as every module needs it)